Give better error messages when registering invalid events
Previously you'd get a `NoMethodError` when you registered an unknown
event but with these changes you'll get a list of each event that was
unknown. Same goes for unhandled events.

This is a breaking change since `MissingEventHandler` is renamed to
`InvalidEventHandlers`, and you wont get `NoMethodError` when
registering an unknown event. That was necessary since it is possible to
have both unknown and unhandled events, in which case I think it is
nicer to just raise one exception. The alternative would be to have two
different exceptions.
These changes should be fine though, I doubt people were rescuing these
exceptions.